+++ lib/channels/channel.dart
@@ -15,9 +15,7 @@ class Channel {
   /// The name of the channel.
   final String name;
 
-  Channel({required this.client, required this.name, bool subscribe = true}) {
-    if (subscribe) this.subscribe();
-  }
+  Channel({required this.client, required this.name});
 
   /// The options of the client.
   PusherOptions get options => client.options;
@@ -30,6 +28,11 @@ class Channel {
   /// Whether the channel is subscribed or not.
   bool get subscribed => _subscribed;
 
+  bool get isPrivate => name.startsWith("private-");
+  bool get isPresence => name.startsWith("presence-");
+  bool get isEncrypted => name.startsWith("private-encrypted-");
+  bool get isPublic => !isPrivate && !isPresence && !isEncrypted;
+
   /// Sets the value of the subscribed property.
   @protected
   set subscribed(bool value) {
@@ -38,7 +41,9 @@ class Channel {
 
   /// Subscribes to the channel.
   void subscribe([bool force = false]) async {
-    if (!client.connected || (subscribed && !force)) {
+    if (!client.connected ||
+        (subscribed && !force) ||
+        client.socketId == null) {
       return;
     }
 
@@ -46,11 +51,9 @@ class Channel {
 
     options.log("SUBSCRIBE", name);
 
-    client.sendEvent("pusher:subscribe", {
-      "channel": name,
-    });
+    client.sendEvent("pusher:subscribe", {"channel": name});
 
-    _subscribed = true;
+    if (isPublic) _subscribed = true;
   }
 
   final _eventsListenersCollection = EventsListenersCollection();
@@ -72,7 +75,10 @@ class Channel {
   /// Unsubscribes from the channel.
   void unsubscribe() {
     client.sendEvent("pusher:unsubscribe", {"channel": name});
+
     _subscribed = false;
+
+    client.channelsCollection.remove(name);
   }
 
   /// Binding for the subscription success event.