+++ CHANGELOG.md
@@ -47,3 +47,8 @@
 ## 0.0.2+6 - 2024-11-01 20:17
 
 - Fix Member add and member remove functions.
+
+## 0.0.2+7 - 2024-11-04 18:16
+
+- Fix channels subscribe & unsubscribe.
+- Fix client reconnect error.